+++ src/test/java/domain/card/CardMachineTest.java
@@ -1,10 +1,8 @@
-package domain;
+package domain.card;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 
-import domain.card.Card;
-import domain.card.CardMachine;
 import exception.ErrorMessage;
 import java.util.HashMap;
 import java.util.Map;
