+++ src/test/java/domain/card/CardTest.java
@@ -1,8 +1,7 @@
-package domain;
+package domain.card;
 
 import constant.Rank;
 import constant.Suit;
-import domain.card.Card;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Nested;
