+++ src/test/java/domain/card/HandTest.java
@@ -1,11 +1,9 @@
-package domain;
+package domain.card;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import constant.Rank;
 import constant.Suit;
-import domain.card.Card;
-import domain.card.Hand;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
