+++ src/test/java/domain/participant/DealerTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.participant;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -6,7 +6,6 @@ import constant.Rank;
 import constant.Suit;
 import domain.card.Card;
 import domain.card.Hand;
-import domain.participant.Dealer;
 import java.util.List;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
