+++ src/test/java/domain/participant/NameTest.java
@@ -1,8 +1,7 @@
-package domain;
+package domain.participant;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 
-import domain.participant.Name;
 import exception.ErrorMessage;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
