+++ src/test/java/domain/participant/NamesTest.java
@@ -1,9 +1,7 @@
-package domain;
+package domain.participant;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 
-import domain.participant.Name;
-import domain.participant.Names;
 import exception.ErrorMessage;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Nested;
